importer/requirements.txt
# attrs==22.2.0
# certifi==2022.12.7
# charset-normalizer==3.0.1
# click==8.1.3
# click-plugins==1.1.1
# cligj==0.7.2
# colorama==0.4.6
# greenlet==2.0.1
# idna==3.4
# munch==2.5.0
# packaging==23.0
# pip==22.3.1
psycopg2==2.9.5
# pyproj==3.4.1
# python-dateutil==2.8.2
python-dotenv==0.21.0
# pytz==2022.7.1
requests==2.28.2
# setuptools==65.6.3
# six==1.16.0
SQLAlchemy==1.4.46
# urllib3==1.26.14
# wheel==0.37.1
# wincertstore==0.2
